Prep for renaming the repo (#6)
