+++ plugin.json
@@ -1,5 +1,5 @@
 {
-    "id": "com.mattermost.cloudapps",
+    "id": "com.mattermost.apps",
     "name": "Cloud Apps",
     "description": "Cloud Apps Registry and API proxy.",
     "version": "0.1.0",
