+++ server/apps/app_get.go
@@ -1,9 +1,9 @@
 // Copyright (c) 2019-present Mattermost, Inc. All Rights Reserved.
 // See License for license information.
 
-package cloudapps
+package apps
 
-import "github.com/mattermost/mattermost-plugin-cloudapps/server/utils"
+import "github.com/mattermost/mattermost-plugin-apps/server/utils"
 
 func (r *registry) GetApp(appID AppID) (*App, error) {
 	if appID != "hello" {
@@ -12,6 +12,6 @@ func (r *registry) GetApp(appID AppID) (*App, error) {
 	return &App{
 		AppID:       "hello",
 		DisplayName: "Hallo სამყარო",
-		RootURL:     "https://levb.ngrok.io/plugin/cloudapps/hello",
+		RootURL:     "https://levb.ngrok.io/plugin/com.mattermost.apps/hello",
 	}, nil
 }
