+++ server/apps/on_user_joined_channel.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2019-present Mattermost, Inc. All Rights Reserved.
 // See License for license information.
 
-package cloudapps
+package apps
 
 import (
 	"github.com/mattermost/mattermost-server/v5/model"
