+++ server/command/info.go
@@ -4,8 +4,8 @@
 package command
 
 import (
-	"github.com/mattermost/mattermost-plugin-cloudapps/server/constants"
-	"github.com/mattermost/mattermost-plugin-cloudapps/server/utils/md"
+	"github.com/mattermost/mattermost-plugin-apps/server/constants"
+	"github.com/mattermost/mattermost-plugin-apps/server/utils/md"
 )
 
 func (c *command) handleInfo(parameters []string) (md.MD, error) {
