+++ server/main.go
@@ -3,8 +3,8 @@ package main
 import (
 	mattermost "github.com/mattermost/mattermost-server/v5/plugin"
 
-	"github.com/mattermost/mattermost-plugin-cloudapps/server/configurator"
-	"github.com/mattermost/mattermost-plugin-cloudapps/server/plugin"
+	"github.com/mattermost/mattermost-plugin-apps/server/configurator"
+	"github.com/mattermost/mattermost-plugin-apps/server/plugin"
 )
 
 var BuildHash string
